requirements.txt
selenium
# 4.9.1
python-dotenv
line-bot-sdk